Split off LandTypeSelectionLoader from LandTypeSelection and create new package
