+++ src/main/java/com/mraof/minestuck/world/lands/gen/LandTypeSelectionLoader.java
@@ -1,4 +1,4 @@
-package com.mraof.minestuck.world.lands;
+package com.mraof.minestuck.world.lands.gen;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
@@ -10,6 +10,7 @@ import com.mojang.serialization.JsonOps;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import com.mraof.minestuck.Minestuck;
 import com.mraof.minestuck.player.EnumAspect;
+import com.mraof.minestuck.world.lands.LandTypes;
 import com.mraof.minestuck.world.lands.terrain.TerrainLandType;
 import com.mraof.minestuck.world.lands.title.TitleLandType;
 import net.minecraft.MethodsReturnNonnullByDefault;
@@ -22,7 +23,6 @@ import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraftforge.event.AddReloadListenerEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.registries.ForgeRegistry;
 import net.minecraftforge.registries.IForgeRegistry;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -31,105 +31,23 @@ import javax.annotation.ParametersAreNonnullByDefault;
 import java.io.IOException;
 import java.io.Reader;
 import java.util.*;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
 @Mod.EventBusSubscriber(modid = Minestuck.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
-public final class LandTypeSelection extends SimplePreparableReloadListener<LandTypeSelection.RawData>
+public final class LandTypeSelectionLoader extends SimplePreparableReloadListener<LandTypeSelectionLoader.RawData>
 {
 	private static final Logger LOGGER = LogManager.getLogger();
 	
 	public static final String TERRAIN_PATH = "minestuck/terrain_land_types.json";
 	public static final String TITLE_PATH = "minestuck/title_land_types.json";
 	
-	private static List<LandsSupplier<TerrainLandType>> terrainGroups;
-	private static Map<EnumAspect, List<LandsSupplier<TitleLandType>>> titleByAspect;
-	
-	public static Collection<List<TerrainLandType>> terrainAlternatives()
-	{
-		return terrainGroups.stream().map(LandsSupplier::get).toList();
-	}
-	
-	public static Collection<List<TitleLandType>> titleAlternatives(EnumAspect aspect)
-	{
-		return titleByAspect.get(aspect).stream().map(LandsSupplier::get).toList();
-	}
-	
-	public static Set<TitleLandType> compatibleTitleTypes(TerrainLandType terrainType)
-	{
-		return titleByAspect.values().stream()
-				.flatMap(list ->
-						list.stream().flatMap(supplier ->
-								supplier.get().stream().filter(titleType ->
-										titleType.isAspectCompatible(terrainType))))
-				.collect(Collectors.toSet());
-	}
-	
-	public static Set<TitleLandType> compatibleTitleTypes(TerrainLandType terrainType, EnumAspect aspect)
-	{
-		return titleByAspect.get(aspect).stream()
-				.flatMap(supplier ->
-						supplier.get().stream().filter(titleType ->
-								titleType.isAspectCompatible(terrainType)))
-				.collect(Collectors.toSet());
-	}
-	
-	public static Set<EnumAspect> compatibleAspects(TitleLandType titleType)
-	{
-		return titleByAspect.entrySet().stream()
-				.flatMap(entry ->
-						entry.getValue().stream().anyMatch(supplier -> supplier.contains(titleType))
-								? Stream.of(entry.getKey())
-								: Stream.empty())
-				.collect(Collectors.toSet());
-	}
-	
-	public interface LandsSupplier<A>
-	{
-		List<A> get();
-		
-		default boolean contains(A element)
-		{
-			return this.get().contains(element);
-		}
-		
-	}
-	
-	private record LandList<A>(List<A> landTypes) implements LandsSupplier<A>
-	{
-		@Override
-		public List<A> get()
-		{
-			return this.landTypes;
-		}
-		
-	}
-	
-	private record LandTag<A>(TagKey<A> tag) implements LandsSupplier<A>
-	{
-		@Override
-		public List<A> get()
-		{
-			ForgeRegistry<A> registry;
-			if(tag.registry().equals(LandTypes.TERRAIN_KEY))
-				registry = (ForgeRegistry<A>) LandTypes.TERRAIN_REGISTRY.get();
-			else if(tag.registry().equals(LandTypes.TITLE_KEY))
-				registry = (ForgeRegistry<A>) LandTypes.TITLE_REGISTRY.get();
-			else
-				throw new RuntimeException("Has tag for unhandled registry: " + this.tag.registry());
-			
-			return Objects.requireNonNull(registry.tags()).getTag(this.tag).stream().toList();
-		}
-	}
-	
 	public record RawData(List<GroupData> terrainTypeData, Map<EnumAspect, List<GroupData>> titleTypeData)
 	{}
 	
 	private record GroupData(Either<List<ResourceLocation>, ExtraCodecs.TagOrElementLocation> value)
 	{
-		private <A> Optional<LandsSupplier<A>> lookup(IForgeRegistry<A> registry)
+		private <A> Optional<LandTypeSelection.LandsSupplier<A>> lookup(IForgeRegistry<A> registry)
 		{
 			return this.value.map(list -> {
 				List<A> elements = new ArrayList<>();
@@ -139,15 +57,15 @@ public final class LandTypeSelection extends SimplePreparableReloadListener<Land
 				if(elements.isEmpty())
 					return Optional.empty();
 				else
-					return Optional.of(new LandList<>(elements));
+					return Optional.of(new LandTypeSelection.LandList<>(elements));
 			}, location -> {
 				if(location.tag())
 				{
 					TagKey<A> tag = TagKey.create(registry.getRegistryKey(), location.id());
-					return Optional.of(new LandTag<>(tag));
+					return Optional.of(new LandTypeSelection.LandTag<>(tag));
 				}
 				else
-					return getOrLog(registry, location.id()).map(landType -> new LandList<>(Collections.singletonList(landType)));
+					return getOrLog(registry, location.id()).map(landType -> new LandTypeSelection.LandList<>(Collections.singletonList(landType)));
 			});
 		}
 		
@@ -167,7 +85,7 @@ public final class LandTypeSelection extends SimplePreparableReloadListener<Land
 	@SubscribeEvent
 	public static void onResourceReload(AddReloadListenerEvent event)
 	{
-		event.addListener(new LandTypeSelection());
+		event.addListener(new LandTypeSelectionLoader());
 	}
 	
 	@Override
@@ -223,19 +141,17 @@ public final class LandTypeSelection extends SimplePreparableReloadListener<Land
 	@Override
 	protected void apply(RawData data, ResourceManager resourceManager, ProfilerFiller profiler)
 	{
-		ImmutableList.Builder<LandsSupplier<TerrainLandType>> terrainGroupsBuilder = ImmutableList.builder();
+		ImmutableList.Builder<LandTypeSelection.LandsSupplier<TerrainLandType>> terrainGroupsBuilder = ImmutableList.builder();
 		
 		for(GroupData terrainGroup : data.terrainTypeData())
 			terrainGroup.lookup(LandTypes.TERRAIN_REGISTRY.get()).ifPresent(terrainGroupsBuilder::add);
 		
-		terrainGroups = terrainGroupsBuilder.build();
 		
-		
-		ImmutableMap.Builder<EnumAspect, List<LandsSupplier<TitleLandType>>> titleMapBuilder = ImmutableMap.builder();
+		ImmutableMap.Builder<EnumAspect, List<LandTypeSelection.LandsSupplier<TitleLandType>>> titleMapBuilder = ImmutableMap.builder();
 		
 		for(EnumAspect aspect : EnumAspect.values())
 		{
-			ImmutableList.Builder<LandsSupplier<TitleLandType>> titleGroupsBuilder = ImmutableList.builder();
+			ImmutableList.Builder<LandTypeSelection.LandsSupplier<TitleLandType>> titleGroupsBuilder = ImmutableList.builder();
 			
 			for(GroupData titleGroup : data.titleTypeData().get(aspect))
 				titleGroup.lookup(LandTypes.TITLE_REGISTRY.get()).ifPresent(titleGroupsBuilder::add);
@@ -243,6 +159,6 @@ public final class LandTypeSelection extends SimplePreparableReloadListener<Land
 			titleMapBuilder.put(aspect, titleGroupsBuilder.build());
 		}
 		
-		titleByAspect = titleMapBuilder.build();
+		LandTypeSelection.setData(terrainGroupsBuilder.build(), titleMapBuilder.build());
 	}
 }
